feat(layers): make Softmax, Linear, Conv1D, VarSelection, CE-loss bf16-instantiable
Generalize the layers and loss a transformer training stack is built from
so they can be instantiated over reduced-precision element types
(float16.BFloat16 in particular), keeping every change generic across
tensor.Numeric -- no model- or workload-specific logic.

- activations.Softmax: relax the type bound from tensor.Float to
  tensor.Numeric. Forward/Backward are expressed entirely via the
  compute.Engine (Softmax/Mul/ReduceSum/Sub); the narrower bound blocked
  bf16 graphs without protecting any invariant.
- core.randomData: convert via ops.FromFloat32 instead of the direct
  T(rand.Float32()) form, which does not compile for the defined-type
  reduced floats (bf16/f16/f8). ops is now a parameter; a nil-ops
  fallback preserves the prior float32/float64 nil-engine construction
  path (Mamba nil-engine guard). Updated callers: NewLinear, NewConv1D,
  NewVariableSelection.
- loss.CrossEntropyLoss: accept targets in any numeric element type by
  converting the class index through numericToFloat64 rather than a
  float32/float64-only type switch, so bf16 targets are usable.

Verified: go build/vet/test green locally on CPU (CUDA-gated GPU tests
skip). Not GPU-verified.

+++ layers/activations/softmax.go
@@ -13,10 +13,15 @@ import (
 
 // Softmax applies the softmax function along a given axis.
 //
-// Softmax is constrained to floating-point element types: the operation is
-// only defined on real-valued inputs, and the backward pass requires the
-// arithmetic implied by tensor.Float.
-type Softmax[T tensor.Float] struct {
+// Softmax is constrained to tensor.Numeric element types. The operation is
+// only meaningful on real-valued inputs, but both forward and backward are
+// expressed entirely through the compute.Engine (Softmax/Mul/ReduceSum/Sub),
+// so the layer is instantiable over any numeric element type the engine
+// supports -- including the reduced-precision floats (float16, bfloat16,
+// float8) used by mixed-precision training. The narrower tensor.Float bound
+// was an over-constraint: it blocked bf16 model graphs without protecting any
+// invariant the body relies on.
+type Softmax[T tensor.Numeric] struct {
 	engine      compute.Engine[T]
 	axis        int
 	outputShape []int
@@ -33,7 +38,7 @@ func (s *Softmax[T]) SetSaver(sv graph.Saver[T]) {
 }
 
 // NewSoftmax creates a new Softmax activation layer.
-func NewSoftmax[T tensor.Float](engine compute.Engine[T], axis int) *Softmax[T] {
+func NewSoftmax[T tensor.Numeric](engine compute.Engine[T], axis int) *Softmax[T] {
 	return &Softmax[T]{engine: engine, axis: axis}
 }
 
@@ -126,7 +131,7 @@ func (s *Softmax[T]) Parameters() []*graph.Parameter[T] { return nil }
 
 // BuildSoftmax constructs a Softmax layer for the registry.
 // The optional "axis" attribute (int or int64) selects the softmax axis; defaults to -1.
-func BuildSoftmax[T tensor.Float](
+func BuildSoftmax[T tensor.Numeric](
 	engine compute.Engine[T],
 	_ numeric.Arithmetic[T],
 	_ string,
